Moved LocationView to its own folder
